snmp.community: move unittests to tests/
